Remove some gcc warnings.
tolower() and friends on some systems pass the argument as the index
to an array and gcc wants that to be an int.

+++ gattrib/src/listsort.c
@@ -96,8 +96,8 @@ int cmp(STRING_LIST *al, STRING_LIST *bl) {
 	  while (isdigit ((int) *b))
 	    b++;
 	}
-      else if (tolower (*a) != tolower (*b))
-	return tolower (*a) - tolower (*b);
+      else if (tolower ((int) *a) != tolower ((int) *b))
+	return tolower ((int) *a) - tolower ((int) *b);
       a++;
       b++;
     }
